Add main thread in WebSocket callback
